Replace miniserde with serde+rmp-serde
Had to do this because miniserde doesn't support enums whose variants
contain data, which I needed for implementing IPC.
I was worried about serde's compile times and stuff, but I was
pleasantly surprised; it's not bad at all! I'd imagine it's only an
issue when you have a ton of complex types you need to derive
serialization for.