Remove the "initially modified" behavior from code
In code it was being applied inconsistently; the only time we got the
"initially modified" status was when using `fs-sentinel` as a library,
since the relevant code path could never be reached through the CLI!
This is because the CLI codepath would validate that only paths which
existed in the current data store were passed in, and this "initially
modified" code path was only run when an unrecognized filesystem
identifier was passed in.

The docs will be updated in the next commit.